backend/requirements.txt
Django==6.0
djangorestframework==3.16.1
django-cors-headers==4.9.0
PyMySQL==1.1.2
beautifulsoup4==4.14.3
requests==2.32.5
sib-api-v3-sdk==7.6.0
python-dotenv==1.2.1

